feat(ras): unify ESP connector strategy for data events
