+++ includes/class-newspack.php
@@ -78,6 +78,11 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/class-plugin-manager.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-theme-manager.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-admin-plugins-screen.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-activation-emails.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-activation.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-data.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-esp-sync.php';
+		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-esp-sync-admin.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-utils.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-data-events.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-webhooks.php';
@@ -86,17 +91,10 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-popups.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-memberships.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/connectors/ga4/class-ga4.php';
-		include_once NEWSPACK_ABSPATH . 'includes/data-events/connectors/class-connector.php';
-		include_once NEWSPACK_ABSPATH . 'includes/data-events/connectors/class-mailchimp.php';
-		include_once NEWSPACK_ABSPATH . 'includes/data-events/connectors/class-activecampaign.php';
+		include_once NEWSPACK_ABSPATH . 'includes/data-events/connectors/class-esp-connector.php';
 		include_once NEWSPACK_ABSPATH . 'includes/data-events/class-woo-user-registration.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-api.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-profile.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-activation-emails.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-activation.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-reader-data.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-esp-sync.php';
-		include_once NEWSPACK_ABSPATH . 'includes/reader-activation/class-esp-sync-admin.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-recaptcha.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-magic-link.php';
 		include_once NEWSPACK_ABSPATH . 'includes/reader-revenue/stripe/class-stripe-connection.php';
